refactor(probability): isolate graph reliability contract
